backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
alembic==1.12.1
pydantic==2.5.2
python-dotenv==1.0.0
pandas==2.1.4
openpyxl==3.1.2
reportlab==4.0.7
python-dateutil==2.8.2
